+++ oap-server/server-fetcher-plugin/kafka-fetcher-plugin/src/main/java/org/apache/skywalking/oap/server/analyzer/agent/kafka/module/KafkaFetcherConfig.java
@@ -43,16 +43,6 @@ public class KafkaFetcherConfig extends ModuleConfig {
     private String groupId = "skywalking-consumer";
 
     /**
-     * Which PartitionId(s) of the topics assign to the OAP server. If more than one, is separated by commas.
-     */
-    private String consumePartitions = "";
-
-    /**
-     * isSharding was true when OAP Server in cluster.
-     */
-    private boolean isSharding = false;
-
-    /**
      * If true, create the Kafka topic when it does not exist.
      */
     private boolean createTopicIfNotExist = true;