+++ oap-server/server-fetcher-plugin/kafka-fetcher-plugin/src/main/java/org/apache/skywalking/oap/server/analyzer/agent/kafka/provider/handler/KafkaHandler.java
@@ -18,39 +18,13 @@
 
 package org.apache.skywalking.oap.server.analyzer.agent.kafka.provider.handler;
 
-import com.google.common.collect.Lists;
-import java.util.Collections;
-import java.util.List;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.utils.Bytes;
-import org.apache.skywalking.oap.server.library.util.StringUtil;
 
 /**
  * A Handler for dealing Message reported by agent. It is binding to a topic of Kafka, and deserialize.
  */
 public interface KafkaHandler {
-
-    /**
-     * Which partition(s) of the topic is handled in cluster mode. Currently, we have to specify the handler working for
-     * partition(s).
-     */
-    default List<TopicPartition> getTopicPartitions() {
-        if (StringUtil.isEmpty(getConsumePartitions())) {
-            return Collections.EMPTY_LIST;
-        }
-
-        List<TopicPartition> topicPartitions = Lists.newArrayList();
-        for (final String partition : getConsumePartitions().trim().split("\\s*,\\s*")) {
-            topicPartitions.add(new TopicPartition(getTopic(), Integer.parseInt(partition)));
-        }
-        return topicPartitions;
-    }
-
-    /**
-     *
-     */
-    String getConsumePartitions();
 
     /**
      * A topic of Kafka is handled.