+++ oap-server/server-fetcher-plugin/kafka-fetcher-plugin/src/main/java/org/apache/skywalking/oap/server/analyzer/agent/kafka/provider/handler/LogHandler.java
@@ -63,11 +63,6 @@ public class LogHandler extends AbstractKafkaHandler {
     }
 
     @Override
-    public String getConsumePartitions() {
-        return config.getConsumePartitions();
-    }
-
-    @Override
     protected String getPlainTopic() {
         return config.getTopicNameOfLogs();
     }